themes update
When you click on bookings, or on filters, there is a section at the top that says '<- bookings' and '<- filters' respectively. i dont know how to change that background. Also there is some fixing to do in regards to the colors (especially that of the text). and i dont know what to do with the primary, secondary and accent color in the white/dark theme